Refactor mesh and light APIs to use std::vector and GLuint
Updated Mesh::CreateMesh to accept std::vector for vertices and indices, improving type safety and flexibility. Changed Light::UseLight parameter types from GLfloat to GLuint for uniform locations. Adjusted related code in main.cpp and other affected files to match new interfaces. Also added initialization for new Shader uniform members and ensured consistent type casting in Window and main.

+++ Baqader_GameEngine/Light.cpp
@@ -12,7 +12,7 @@ Light::Light(GLfloat red, GLfloat green, GLfloat blue, GLfloat ambientIntensity)
 	m_ambientIntensity = ambientIntensity;
 }
 
-void Light::UseLight(GLfloat ambientIntensityLocation, GLfloat ambientColourLocation)
+void Light::UseLight(GLuint ambientIntensityLocation, GLuint ambientColourLocation)
 {
 	glUniform3f(ambientColourLocation, m_colour.x, m_colour.y, m_colour.z);
 	glUniform1f(ambientIntensityLocation, m_ambientIntensity);
